library search: Swap precedence of /usr{/local,}/lib
It is customary for /usr/local to contain software installed by the
local system administrator/owner and for that to override the same
software if provided at a system-wide level (e.g. in /usr/lib)

This addresses one of the issues touched on in #61 but does not fix
the key issue (IMHO) which is that `make install PREFIX=$HOME`
does not result in a working `git issue` so not marking it as fixed
in these commits.